src/main/java/dang/aishwarya/service/VehicleServiceImpl.java
package dang.aishwarya.service;

import dang.aishwarya.entity.Vehicle;
import dang.aishwarya.exception.VehicleNotCreatedException;
import dang.aishwarya.repository.VehicleRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;
import java.util.Optional;

@Service
public class VehicleServiceImpl implements VehicleService {

    @Autowired
    VehicleRepository vehicleRepository;

    @Override
    public List<Vehicle> findAll() {
        return (List<Vehicle>) vehicleRepository.findAll();
    }

    @Override
    public void findOne(String vin) {
    }


    @Override
    @Transactional
    public Vehicle create(Vehicle vehicle) {
        return vehicleRepository.save(vehicle);
    }
}
